ml/requirements.txt
Flask 
Flask-Cors 
tensorflow 
numpy 
opencv-python 
dlib 
face-recognition 
pillow 
scipy 
gunicorn 
psutil 
